Asset... Assets... fucking get out of here wiith this
