+++ src/core/gui_core_headers.h
@@ -63,7 +63,7 @@
 #include "../gui/CTF1DPanel.h"
 #include "../gui/MainFrame.h"
 #include "../gui/ErrorDialog.h"
-#include "../gui/MyAssetPanelParent.h"
+#include "../gui/AssetObjectPanel.h"
 #include "../gui/MyMovieAssetPanel.h"
 #include "../gui/MyImageAssetPanel.h"
 #include "../gui/MyParticlePositionAssetPanel.h"
@@ -89,7 +89,7 @@
 #include "../gui/MyResultsPanel.h"
 #include "../gui/ActionsPanelSpa.h"
 #include "../gui/ActionsPanelTm.h"
-#include "../gui/MyAssetsPanel.h"
+#include "../gui/AssetsPanel.h"
 #include "../gui/MySettingsPanel.h"
 #include "../gui/post_PickingResultsPanel.h"
 #include "../gui/MyParticlePositionExportDialog.h"
